* Changed physical_entry::{address,size} to phys_{addr,size}_t and changed   map_physical_memory()'s physicalAddress parameter type from void* to   phys_addr_t. This breaks source compatibility, but -- as long as   phys_{addr,size}_t remain 32 bit wide -- keeps binary compatibility with   BeOS. * Adjusted all code using the affected interfaces (Oh what fun!). Added a few   TODOs in places where the wrong types (e.g. void* for physical addresses   are used). Looks like quite a few drivers aren't 64 bit safe and others   will break with PAE.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@36960 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/drivers/network/ipro1000/if_em_osdep.h
@@ -125,7 +125,7 @@ static inline unsigned long vtophys(unsigned long virtual_addr)
 	err = get_memory_map((void *)virtual_addr, 2046, &pe, 1);
 	if (err < 0)
 		panic("ipro1000: get_memory_map failed for %p, error %08lx\n", (void *)virtual_addr, err);
-	return (unsigned long) pe.address;
+	return pe.address;
 }
 
 #define M_DEVBUF 1
